Refactor codegen model classes
