+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/compiler/bir/codegen/model/JavaField.java
@@ -17,8 +17,8 @@
  */
 package org.wso2.ballerinalang.compiler.bir.codegen.model;
 
-import org.wso2.ballerinalang.compiler.bir.codegen.JvmCodeGenUtil;
 import org.wso2.ballerinalang.compiler.bir.codegen.interop.JFieldMethod;
+import org.wso2.ballerinalang.compiler.bir.codegen.utils.JvmCodeGenUtil;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
@@ -34,38 +34,31 @@ public class JavaField {
     private final Field field;
 
     public JavaField(JFieldMethod method, Field field) {
-
         this.method = method;
         this.field = field;
     }
 
     public String getDeclaringClassName() {
-
         return field.getDeclaringClass().getName().replace(".", "/");
     }
 
     public String getName() {
-
         return field.getName();
     }
 
     public boolean isStatic() {
-
         return Modifier.isStatic(field.getModifiers());
     }
 
     public JFieldMethod getMethod() {
-
         return method;
     }
 
     public String getSignature() {
-
         return JvmCodeGenUtil.getSig(field.getType());
     }
 
     public Class<?> getFieldType() {
-
         return field.getType();
     }
 }
